requirements.txt
adal==1.2.4
azure-common==1.1.25
azure-core==1.6.0
azure-servicebus==7.0.0b3
certifi==2020.4.5.2
cffi==1.14.0
chardet==3.0.4
click==7.1.2
cryptography==2.9.2
idna==2.9
isodate==0.6.0
msrest==0.6.16
msrestazure==0.6.3
oauthlib==3.1.0
pycparser==2.20
PyJWT==1.7.1
python-dateutil==2.8.1
requests==2.23.0
requests-oauthlib==1.3.0
six==1.15.0
uamqp==1.2.8
urllib3==1.25.9
